group comparisons moved to new file
comparegroups.py
